refactor(EnumInterop+Hashable.swift): Dedicated file for comms type Hashable conformance
